requirements.txt
flask==3.0.0
openai-whisper==20231117
torch
ffmpeg-python
werkzeug==3.0.1
gunicorn==21.2.0
